+++ catalogue/OracleCatalogue.cpp
@@ -855,14 +855,14 @@ void OracleCatalogue::deleteArchiveFile(const std::string &diskInstanceName, con
          .add("selectFromArchiveFileTime", selectFromArchiveFileTime);
       for(auto it=archiveFile->tapeFiles.begin(); it!=archiveFile->tapeFiles.end(); it++) {
         std::stringstream tapeCopyLogStream;
-        tapeCopyLogStream << "copy number: " << it->copyNb
+        tapeCopyLogStream << "copy number: " << static_cast<int>(it->copyNb)
           << " vid: " << it->vid
           << " fSeq: " << it->fSeq
           << " blockId: " << it->blockId
           << " creationTime: " << it->creationTime
           << " fileSize: " << it->fileSize
           << " checksumBlob: " << it->checksumBlob //this shouldn't be here: repeated field
-          << " copyNb: " << it->copyNb //this shouldn't be here: repeated field
+          << " copyNb: " << static_cast<int>(it->copyNb) //this shouldn't be here: repeated field
           << " supersededByVid: " << it->supersededByVid
           << " supersededByFSeq: " << it->supersededByFSeq;
         spc.add("TAPE FILE", tapeCopyLogStream.str());
@@ -929,7 +929,7 @@ void OracleCatalogue::deleteArchiveFile(const std::string &diskInstanceName, con
         << " creationTime: " << it->creationTime
         << " fileSize: " << it->fileSize
         << " checksumBlob: " << it->checksumBlob //this shouldn't be here: repeated field
-        << " copyNb: " << it->copyNb //this shouldn't be here: repeated field
+        << " copyNb: " << static_cast<int>(it->copyNb) //this shouldn't be here: repeated field
         << " supersededByVid: " << it->supersededByVid
         << " supersededByFSeq: " << it->supersededByFSeq;
       spc.add("TAPE FILE", tapeCopyLogStream.str());
